2: size_t instead of int

+++ 2-selection_sort.c
@@ -22,7 +22,7 @@ void swap(int *x, int *y)
  */
 void selection_sort(int *array, size_t size)
 {
-	int i, j, min;
+	size_t i, j, min;
 
 	if (!array || size < 2)
 		return;
